+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Policy scoped grade v2 API
+
 ## 0.16.0 - Search
 - Updated table styles to simplify and unify table margins
 - Implemented Search for Students table